Add files via upload
Here Includes Create Reactive Forms, Add Validation, and Duplicate the element. So here, we are not connected to the server. So In Part 2 We have created new application and connected it with the Server. In here Included Word file. that about Create a Reactive Form. A to Z. So refer it. You can get a good idea about reactive forms.